Prevent starving fibers awaiting watchers (close #7)
This change adds an algorithm to prevent starving the ev_loop when
there's always runnable fibers (as is the case with `snooze`). It works
in the following manner (look at `Thread_schedule_fiber()`):

- `next_fiber` is shifted from the run queue
- if `next_fiber` is not nil and the ref_count is greater than 0
  - a `run_no_wait_count` associated with the selector is incremented
  - return if `run_no_wait_count` is smaller than runnable fiber count
  - return if `run_no_wait_count` is smaller than 10
  - reset `run_no_wait_count`
  - run the ev_loop with `EV_NOWAIT` meaning the loop does not wait
    for events and returns after a single iteration

So basically the ev_loop is guaranteed to run at least every `x`
switchpoints where `x` is the current runnable fiber count, provided the
ref_count (number of waiting fibers) is greater than 0.