Implemented child fibers
This adds support for the structured concurrency model of nested fibers.
Fibers are spun always in the context of a containing fiber, resulting
in a tree-like structure of fiber *ownership*.

This change necessitated a lot of changes to the way threads are
handled. For example, `Thread#raise` and `Thread#kill` were patched to
schedule errors on the thread's main fiber, instead of the stock
implementation which could interrupt the currently executing fiber,
which could happen also while the thread's event loop is running.

The last point is also relevant for the main fiber, specifically the
handling of signals. For that effect, Polyphony installs signal traps
for `SIGINT` and `SIGTERM`, making sure the signal handling occurs out
of the event loop.