Fix possible race condition in Gyro::Queue
A race condition could happen where a fiber waits on a queue and is
interrupted. Before, when `#push` was called on a queue with waiting
fibers, the corresponding async watcher would have been signalled with
the pushed value, without actually putting it on the queue. If the
waiting fiber was interrupted before running again, the pushed value
would have been lost. This change ensures the value is *always* added to
the queue, and that the fiber calling `#shift` waits to be signalled,
then actually shifts the value from the queue.