Fix forking behaviour
The interaction of multi-fiber programs and forking can lead to
undefined behaviour, and if not carefully controlled will cause a
`FiberError` "fiber called across stack rewinding barrier" exception.

If a fork is done from a fiber other than the main fiber, the currently
active fiber becomes the main fiber of the forked process. In addition,
when the forked process exits, the new main fiber will attempt a
transfer back to some other fiber that transferred control to it, once
it is terminated. To make matters worse, any at_exit handler will be ran
using the current fiber at the time it was defined, which can cause a
big mess.

Our solution consists of the following:

- On fork, create a new fiber using `Fiber.new` and transfer control to
  it, making it the new main fiber.
- Make sure the program is terminated while running the new main fiber,
  i.e. rescue *all* exceptions and call `exit` and `exit!` explicitly.
- Use the `at_exit` handler that kills the main fiber's child fibers
  only on the non-forked process.