Replace core Queue class with Polyphony::Queue (close #22)
The issue: `concurrent-ruby` 1.1.6 includes an implementation of thread
local vars that starts a thread used to finalize thread local vars. This
thread pops items from a Queue instance. The core Ruby Queue class is of
course not fiber-aware. Polyphony, on process exit, kills and joins all
threads, leading to the program hanging.

Our solution, since there's going to have other problems like this when
integrating Polyphony with other programs, is to replace the core Queue
class with the system agent-provided implementation.