Simplify Queue shift queue handling
For better performance when waiting on events, we introduce a new
`LibevAgent_wait_event` API that creates an async watcher and waits for
it to be signalled. The async watcher will in fact never be signalled.
Instead, another fiber will schedule the waiting fiber. This provides
a way to avoid spinning the CPU when waiting for an event when no work
is left to do. This change to LibevQueue makes it completely
independent of libev (that is, it can work with any agent), and also
provides a nice performance boost, around 3-4% in the simple http
server example.